+++ bin/store/CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## 0.2.60
+### Author
+Bert
+
+### Added
+  - ***Prometheus metrics***:
+    - `METRICS_QUERIES.md`
+      - Add Prometheus metrics queries route for API and system metrics.
+      - Add Prometheus metrics results route for API and system metrics.
+
+## 0.2.60
 
 ### Author
 Jean